- Schedule notification changes
