+++ src/pages/Home.tsx
@@ -100,7 +100,7 @@ import RegularAccount from '../bitcoin/services/accounts/RegularAccount';
 import GetBittrRecurringBuyContents from './GetBittr/GetBittrRecurringBuyContent';
 import firebase from 'react-native-firebase';
 import NotificationListContent from '../components/NotificationListContent';
-import { firebaseNotificationListener } from '../common/CommonFunctions/notifications';
+import { firebaseNotificationListener, scheduleNotification } from '../common/CommonFunctions/notifications';
 // const { Value, abs, sub, min } = Animated
 // const snapPoints = [ Dimensions.get( 'screen' ).height - 150, 150 ]
 // const position = new Value( 1 )
@@ -524,13 +524,7 @@ export default function Home(props) {
   useEffect(function () {
     firebaseNotificationListener();
     storeFCMToken();
-    (async () => {
-      const enabled = await firebase.messaging().hasPermission();
-      if (enabled) {
-        scheduleNotification();
-      }
-    })();
-
+    
     let focusListener = props.navigation.addListener('didFocus', () => {
       setCurrencyCodeFromAsync();
       getAssociatedContact();
@@ -571,38 +565,7 @@ export default function Home(props) {
   //   return unsubscribe;
   // }, []);
 
-  const scheduleNotification = async () => {
-    const notification = new firebase.notifications.Notification()
-      .setTitle('abcd')
-      .setBody('xyz')
-      .setNotificationId('1')
-      .setSound('default')
-      .setSubtitle('www')
-      .setData({ title: 'qqq', body: 'rrrr' })
-      .android.setChannelId('reminder')
-      .android.setPriority(firebase.notifications.Android.Priority.High);
-
-    // Schedule the notification for 1 minute in the future
-    const date = new Date();
-    date.setSeconds(date.getSeconds() + 20);
-    // date.setMinutes(date.getMinutes() + 1);
-
-    await firebase
-      .notifications()
-      .scheduleNotification(notification, {
-        fireDate: date.getTime(),
-        repeatInterval: 'minute',
-      })
-      .then(() => {})
-      .catch((err) => console.log('err', err));
-    firebase
-      .notifications()
-      .getScheduledNotifications()
-      .then((notifications) => {
-        console.log('logging notifications', notifications);
-      });
-  };
-
+  
   const storeFCMToken = async () => {
     const fcmToken = await firebase.messaging().getToken();
     console.log('fcmToken', fcmToken);
@@ -611,7 +574,6 @@ export default function Home(props) {
       dispatch(updateFCMTokens(fcmArray));
     }
   };
-
   const setCurrencyCodeFromAsync = async () => {
     let currencyCodeTmp = await AsyncStorage.getItem('currencyCode');
     if (!currencyCodeTmp) {
